Make progress fixing reference tests
* Fix SpecConfigMerge so it implements SpecConfigAltair and toVersionAltair() still returns the merge instance instead of unwrapping
* Support sync aggregate in merge block processing
