+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/config/SpecConfigAltair.java
@@ -13,59 +13,13 @@
 
 package tech.pegasys.teku.spec.config;
 
-import java.util.Objects;
 import java.util.Optional;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.ssz.type.Bytes4;
 
-public class SpecConfigAltair extends DelegatingSpecConfig {
+public interface SpecConfigAltair extends SpecConfig {
 
-  // Updated penalties
-  private final UInt64 inactivityPenaltyQuotientAltair;
-  private final int minSlashingPenaltyQuotientAltair;
-  private final int proportionalSlashingMultiplierAltair;
-
-  // Misc
-  private final int syncCommitteeSize;
-  private final UInt64 inactivityScoreBias;
-  private final UInt64 inactivityScoreRecoveryRate;
-
-  // Time
-  private final int epochsPerSyncCommitteePeriod;
-
-  // Fork
-  private final Bytes4 altairForkVersion;
-  private final UInt64 altairForkEpoch;
-
-  // Sync protocol
-  private final int minSyncCommitteeParticipants;
-
-  public SpecConfigAltair(
-      final SpecConfig specConfig,
-      final UInt64 inactivityPenaltyQuotientAltair,
-      final int altairMinSlashingPenaltyQuotient,
-      final int proportionalSlashingMultiplierAltair,
-      final int syncCommitteeSize,
-      final UInt64 inactivityScoreBias,
-      final UInt64 inactivityScoreRecoveryRate,
-      final int epochsPerSyncCommitteePeriod,
-      final Bytes4 altairForkVersion,
-      final UInt64 altairForkEpoch,
-      final int minSyncCommitteeParticipants) {
-    super(specConfig);
-    this.inactivityPenaltyQuotientAltair = inactivityPenaltyQuotientAltair;
-    this.minSlashingPenaltyQuotientAltair = altairMinSlashingPenaltyQuotient;
-    this.proportionalSlashingMultiplierAltair = proportionalSlashingMultiplierAltair;
-    this.syncCommitteeSize = syncCommitteeSize;
-    this.inactivityScoreBias = inactivityScoreBias;
-    this.inactivityScoreRecoveryRate = inactivityScoreRecoveryRate;
-    this.epochsPerSyncCommitteePeriod = epochsPerSyncCommitteePeriod;
-    this.altairForkVersion = altairForkVersion;
-    this.altairForkEpoch = altairForkEpoch;
-    this.minSyncCommitteeParticipants = minSyncCommitteeParticipants;
-  }
-
-  public static SpecConfigAltair required(final SpecConfig specConfig) {
+  static SpecConfigAltair required(SpecConfig specConfig) {
     return specConfig
         .toVersionAltair()
         .orElseThrow(
@@ -75,82 +29,26 @@ public class SpecConfigAltair extends DelegatingSpecConfig {
                         + specConfig.getClass().getSimpleName()));
   }
 
-  public Bytes4 getAltairForkVersion() {
-    return altairForkVersion;
-  }
+  Bytes4 getAltairForkVersion();
 
-  public UInt64 getAltairForkEpoch() {
-    return altairForkEpoch;
-  }
+  UInt64 getAltairForkEpoch();
 
-  public UInt64 getInactivityPenaltyQuotientAltair() {
-    return inactivityPenaltyQuotientAltair;
-  }
+  UInt64 getInactivityPenaltyQuotientAltair();
 
-  public int getMinSlashingPenaltyQuotientAltair() {
-    return minSlashingPenaltyQuotientAltair;
-  }
+  int getMinSlashingPenaltyQuotientAltair();
 
-  public int getProportionalSlashingMultiplierAltair() {
-    return proportionalSlashingMultiplierAltair;
-  }
+  int getProportionalSlashingMultiplierAltair();
 
-  public int getSyncCommitteeSize() {
-    return syncCommitteeSize;
-  }
+  int getSyncCommitteeSize();
 
-  public UInt64 getInactivityScoreBias() {
-    return inactivityScoreBias;
-  }
+  UInt64 getInactivityScoreBias();
 
-  public UInt64 getInactivityScoreRecoveryRate() {
-    return inactivityScoreRecoveryRate;
-  }
+  UInt64 getInactivityScoreRecoveryRate();
 
-  public int getEpochsPerSyncCommitteePeriod() {
-    return epochsPerSyncCommitteePeriod;
-  }
+  int getEpochsPerSyncCommitteePeriod();
 
-  public int getMinSyncCommitteeParticipants() {
-    return minSyncCommitteeParticipants;
-  }
+  int getMinSyncCommitteeParticipants();
 
   @Override
-  public Optional<SpecConfigAltair> toVersionAltair() {
-    return Optional.of(this);
-  }
-
-  @Override
-  public boolean equals(final Object o) {
-    if (this == o) return true;
-    if (o == null || getClass() != o.getClass()) return false;
-    final SpecConfigAltair that = (SpecConfigAltair) o;
-    return Objects.equals(specConfig, that.specConfig)
-        && minSlashingPenaltyQuotientAltair == that.minSlashingPenaltyQuotientAltair
-        && proportionalSlashingMultiplierAltair == that.proportionalSlashingMultiplierAltair
-        && syncCommitteeSize == that.syncCommitteeSize
-        && Objects.equals(inactivityScoreBias, that.inactivityScoreBias)
-        && Objects.equals(inactivityScoreRecoveryRate, that.inactivityScoreRecoveryRate)
-        && epochsPerSyncCommitteePeriod == that.epochsPerSyncCommitteePeriod
-        && minSyncCommitteeParticipants == that.minSyncCommitteeParticipants
-        && Objects.equals(inactivityPenaltyQuotientAltair, that.inactivityPenaltyQuotientAltair)
-        && Objects.equals(altairForkVersion, that.altairForkVersion)
-        && Objects.equals(altairForkEpoch, that.altairForkEpoch);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(
-        specConfig,
-        inactivityPenaltyQuotientAltair,
-        minSlashingPenaltyQuotientAltair,
-        proportionalSlashingMultiplierAltair,
-        syncCommitteeSize,
-        inactivityScoreBias,
-        inactivityScoreRecoveryRate,
-        epochsPerSyncCommitteePeriod,
-        altairForkVersion,
-        altairForkEpoch,
-        minSyncCommitteeParticipants);
-  }
+  Optional<SpecConfigAltair> toVersionAltair();
 }
